fix (tests): More changes due to the element-registry refactoring
