+++ platform/nativescript/compiler.js
@@ -1,3 +1,3 @@
 export { parseComponent } from './compiler/sfc/parser'
 export { compile, compileToFunctions } from './compiler/index'
-export { registerElement } from './element-registry'
+export { registerElement } from './register'
